TheDirectory_v8_regenerated - Copy/requirements.txt

Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
requests==2.32.3
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-Admin==2.5.1
Flask-Session==0.5.0
Flask-Session==0.5.0
Flask-Session==0.5.0

